Start creating loaders

+++ src/Repetitorg/Storage/DatabaseRawEntities/ClientEntity.cs
@@ -6,7 +6,7 @@ namespace Storage.SQLite.DatabaseRawEntities
 {
     class ClientEntity : EntityWithId
     {
-        public int BalanceInCopex { get; set; }
+        public long BalanceInCopex { get; set; }
         public long PersonDataId { get; set; }
     }
 }
